🚚 break out core realtime server types
