+++ packages/atom.io/src/realtime-server/index.ts
@@ -1,6 +1,3 @@
-import type { RootStore } from "atom.io/internal"
-import type { Json } from "atom.io/json"
-
 export * from "./continuity/prepare-to-sync-realtime-continuity"
 export * from "./ipc-sockets"
 export * from "./realtime-action-receiver"
@@ -10,20 +7,5 @@ export * from "./realtime-mutable-provider"
 export * from "./realtime-server-stores"
 export * from "./realtime-state-provider"
 export * from "./realtime-state-receiver"
-
-export type Socket = {
-	id: string | undefined
-	on: (event: string, listener: (...args: Json.Serializable[]) => void) => void
-	onAny: (
-		listener: (event: string, ...args: Json.Serializable[]) => void,
-	) => void
-	off: (event: string, listener: (...args: Json.Serializable[]) => void) => void
-	offAny: (
-		listener: (event: string, ...args: Json.Serializable[]) => void,
-	) => void
-	emit: (event: string, ...args: Json.Serializable[]) => void
-}
-export type ServerConfig = {
-	socket: Socket
-	store?: RootStore
-}
+export type * from "./server-config"
+export type * from "./socket-interface"
