+++ packages/atom.io/src/realtime-server/server-config.ts
@@ -1,0 +1,8 @@
+import type { RootStore } from "atom.io/internal"
+
+import type { Socket } from "./socket-interface"
+
+export type ServerConfig = {
+	socket: Socket
+	store?: RootStore
+}
